requirements.txt
-e git+http://github.com/mitsuhiko/flask.git#egg=Flask-dev
Flask-HTTPAuth==2.3.0
Flask-RESTful==0.3.1
Flask-SQLAlchemy==2.0
Jinja2==2.7.3
MarkupSafe==0.23
SQLAlchemy==0.9.8
Werkzeug==0.10.1
aniso8601==0.92
argparse==1.2.1
blinker==1.3
click==3.3
itsdangerous==0.24
names==0.3.0
passlib==1.6.2
psycopg2==2.6
pytz==2014.10
six==1.9.0
wsgiref==0.1.2
